fix(psnr): widen the noise multiply to double (CodeQL 1009) (#1361)
* fix(psnr): widen the noise multiply to double (CodeQL 1009)

`compute_psnr()` accumulated `diff * diff` in `float` before widening to
its `double` accumulator, which CodeQL reports as
cpp/integer-multiplication-cast-to-long ("Multiplication result may
overflow 'float' before it is converted to 'double'"). One operand is now
widened so the multiply happens in double.

Verified rather than assumed that this moves nothing: compute_psnr() has
no call sites anywhere in the tree -- nm on liblibvmaf_feature.a shows
T compute_psnr with no matching U, and the live PSNR paths are
integer_psnr.c and float_psnr.c -- and it has no SIMD twin. All three
Netflix golden pairs are bit-identical before and after at
--precision=max (0 differing keys across every metric) and
meson test --suite=fast is unchanged.

The sibling alert 1005 on core/src/feature/iqa/convolve.c:155 is
deliberately NOT fixed. The identical change to iqa_convolve's four
accumulation sites fails the build's own bit-exactness test
(test_iqa_convolve: test_gauss_12x12, "avx2 convolve output not
bit-identical to scalar"; control is 13/13 at the same commit). That is
the invariant ADR-0138 exists to protect -- the AVX2 twin multiplies in
float and widens after, via _mm256_mul_ps then _mm256_cvtps_pd,
specifically to mirror the scalar. Widening the scalar desynchronises
AVX2, AVX-512 and NEON at once and would move every SSIM / MS-SSIM score,
for an overflow the input domain forbids: image samples times a
normalised kernel give |a*b| <= 255, ~38 orders of magnitude below
FLT_MAX. Recorded in research digest 2031; per the standing rule the
alert stays open for the maintainer to decide.

Also verified while here: epic #1243's remaining checkbox
(ai/scripts/extract_ugc_features.py:175) is already satisfied on master.


* docs(state): cite PR #1361 instead of the placeholder (ADR-0165 touch gate)

---------

+++ core/src/feature/psnr.c
@@ -40,7 +40,7 @@ int compute_psnr(const float *ref, const float *dis, int w, int h, int ref_strid
             float ref_ = ref[(ptrdiff_t)i * ref_stride_ + j];
             float dis_ = dis[(ptrdiff_t)i * dis_stride_ + j];
             float diff = ref_ - dis_;
-            noise_ += diff * diff;
+            noise_ += (double)diff * diff;
         }
     }
     noise_ /= ((double)w * h);
